✨ feat(ios): migrate iOS interop from SKIE to native Swift Export (#789)
* ✨ feat(ios): complete SKIE → Swift Export cutover

* 🐛 fix(ios): restore test-target compile after Swift Export cutover

+++ iosApp/ListenUp/Features/LibrarySetup/LibraryReadinessObserver.swift
@@ -67,7 +67,7 @@ struct ScanProgress: Equatable {
             isPersisting: state.phase == "persisting",
             filesDone: Int(state.current),
             filesTotal: Int(state.filesTotal),
-            fraction: state.progressFraction.map { Double($0.floatValue) },
+            fraction: state.progressFraction.map { Double($0) },
             books: Int(state.books),
             authors: Int(state.authors),
             hours: Int(state.hours),
